+++ examples/axum/src/graphql/model/Droid.rs
@@ -22,7 +22,7 @@ impl Droid {
         self.name.clone()
     }
 
-    pub async fn friends(&self, first: Option<i64>, after: Option<ID>) -> FriendsConnection {
+    pub async fn friends(&self, first: Option<i32>, after: Option<ID>) -> FriendsConnection {
         FriendsConnection {
             totalCount: Some(4),
             edges: vec![
